fix: danmaku_settings_sheet.dart int→double 编译错误
_closestIntIndex 里 (steps[i] - v).abs() 返回 int, 赋给 double
bestDist 报错. 加 .toDouble().

+++ lib/danmaku/widgets/danmaku_settings_sheet.dart
@@ -145,7 +145,7 @@ class _DanmakuSettingsSheetState extends State<DanmakuSettingsSheet> {
     var best = 0;
     var bestDist = double.infinity;
     for (var i = 0; i < steps.length; i++) {
-      final d = (steps[i] - v).abs();
+      final d = (steps[i] - v).abs().toDouble();
       if (d < bestDist) {
         bestDist = d;
         best = i;
